requirements.txt
Flask
itsdangerous
pyjwt
Flask-Mail
flask_cors
flask_mongoengine
dateparser